requirements.txt
python-telegram-bot==22.3
APScheduler==3.10.4
anyio==4.10.0
certifi==2024.12.14
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pytz==2025.2
six==1.17.0
sniffio==1.3.1
tzdata==2025.2
tzlocal==5.3.1
